requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers==0.30.3
transformers==4.45.2
accelerate==0.34.0
safetensors>=0.3.1
Pillow>=9.5.0
invisible-watermark>=0.2.0
peft>=0.4.0
opencv-python>=4.8.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9
